angler: fix first audio track
It seems that the first time an audio track was played was when the
audio clibration data is sent down to the hw. However it wasn't being
properly applied or we missed it and that first track will always sound
awful.

The speaker protection extension fixes this by pre-loading the
calibration data before the first track is played.

Ticket: OSS CYAN-7486
Change-Id: Ibb1ace9dfbf2c1dec4844d1eedd8843889415792
Signed-off-by: Roman Birg <roman@cyngn.com>